Moved AC Automaton to a separate file.
